Projects can be deleted
The projects menu can become cluttered with half-written songs that never
went anywhere etc. and now there is a way to clean it up. This will be
especially useful in the future should a "feed" function be implemented,
which might continually queue up old projects at random for you to
continue working on. Additionally, this functionality can be used to
abort any song in progress, rather than having to wait forit to play in full
and then export.