Adjustable loop points
Previously there was no way to manage a particular section of the song by
itself without playing through the song in its entirety. Now any portion
can be set to loop, allowing the user to record in just that area or to set
an unchanging backing track to work on top of.