. hotkey melds comp into spans
Previously sections of past recorded material could be auditioned, but
could not strictly be added into the song, only used as backing loops or
included in the final recording by letting the whole song play out and
export. Now they will be melded into the playing spans as one larger span,
so the user may continue their work as usual.