[v 1.13.1] OpenGL timeline
Spans had been displayed using scripting to create and erase `panel`s with
every change, which was very ineligant and beginning to cause lag for being
so heavyweight. Now OpenGL is used in `js` with a very lightweight
implementation, which is the way it should have been done all along.