NEW GUI FOR VISUALIZATION OF THE RHYTHMIC ANALYSIS
Recently The Amanuensis got an intelligence upgrade with a complete rework
of the section of code handling its rhythmic analysis, colloquially called
its "consciousness". I describe here another upgrade featuring an in-depth
presentation-view visualization of the analysis as it's being conducted,
which essentially shows the player a scoring of what they're playing.
In addition, it will be important in understanding how well the analysis
itself is working and giving clues as to how it might be improved in the
future.